requirements.txt
h5py==3.6.0
ipympl==0.8.8
ipywidgets==7.6.5
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
Ninja==1.10.2.3
numpy==1.22.2
pandas==1.4.0
Pillow==9.2.0
pytorch-lightning==1.6.5
scipy==1.8.0
seaborn==0.11.2
torch==1.13.1
torch-ema==0.3
torchaudio==0.13.1
torchvision==0.14.1
torchinfo==1.6.3
tqdm==4.63.0
wandb==0.12.11
piq==0.7.1
